refactor: movieThumbnail 파일 분리
