+++ src/components/MoviePreview/createMoviePreview.ts
@@ -1,6 +1,6 @@
 import "./style.css";
 
-import MovieThumbnail from "./createMovieThumbnail";
+import MovieThumbnail from "../MovieThumbnail/createMovieThumbnail";
 import createElement from "../../utils/createElement";
 import starFills from "./star_filled.png";
